+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# 1.4.0
+# 2.1.0
 
-```diff
-+ Added Database.createCollection
+```ts
++ Added selectWithElementMatch - Matches a deeper nested value in an array.
++ Added doesCollectionExist
++ Added argument 'returnFalseIfExists' to createCollection to return false if collection already exists.
 ```